+++ src/plugins/Plugin.ts
@@ -1,14 +1,15 @@
 import { commands } from 'vscode';
 import MarkdownIt from 'markdown-it';
+import { camel2Kebab } from './utils'
 import { config } from '../config';
 
 export abstract class Plugin {
   private _name: string;
   private _package: string;
 
-  constructor(name: string, packageName: string) {
+  constructor(name: string) {
     this._name = name;
-    this._package = packageName;
+    this._package = `markdown-it-${camel2Kebab(name)}`
 
     config.on([`markdownEnhanced.${this._name}`], () => {
       commands.executeCommand('markdown.api.reloadPlugins');
@@ -22,8 +23,11 @@ export abstract class Plugin {
   load(md: MarkdownIt) {
     if (this.enable) {
       md.use(require(this._package), this.options);
+      this.extra(md)
     }
   }
 
   abstract get options(): Object;
+
+  extra(md: MarkdownIt) {}
 }