+++ src/plugins/index.ts
@@ -1,9 +1,25 @@
+import { MarkdownItAbbr } from './MarkdownItAbbr';
 import { MarkdownItAnchor } from './MarkdownItAnchor';
+import { MarkdownItAttrs } from './MarkdownItAttrs';
+import { MarkdownItDeflist } from './MarkdownItDeflist';
 import { MarkdownItFootnote } from './MarkdownItFootnote';
+import { MarkdownItMark } from './MarkdownItMark'
+import { MarkdownItSub } from './MarkdownItSub';
+import { MarkdownItSup } from './MarkdownItSup';
 import { MarkdownItTableOfContents } from './MarkdownItTableOfContents';
+import { MarkdownItTaskLists } from './MarkdownItTaskLists';
+import { MarkdownItUnderline } from './MarkdownItUnderline';
 
 export const plugins = [
+  MarkdownItAbbr,
   MarkdownItAnchor,
+  MarkdownItAttrs,
+  MarkdownItDeflist,
   MarkdownItFootnote,
+  MarkdownItMark,
+  MarkdownItSub,
+  MarkdownItSup,
   MarkdownItTableOfContents,
+  MarkdownItTaskLists,
+  MarkdownItUnderline,
 ];