+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## v0.3.0
+
+- [markdown-it-abbr](https://github.com/markdown-it/markdown-it-abbr)
+- [markdown-it-attrs](https://github.com/arve0/markdown-it-attrs)
+- [markdown-it-deflist](https://github.com/markdown-it/markdown-it-deflist)
+- [markdown-it-mark](https://github.com/markdown-it/markdown-it-mark)
+- [markdown-it-sub](https://github.com/markdown-it/markdown-it-sub)
+- [markdown-it-sup](https://github.com/markdown-it/markdown-it-sup)
+- [markdown-it-task-lists](https://github.com/revin/markdown-it-task-lists)
+- [markdown-it-underline](https://github.com/arve0/markdown-it-underline)
+- Fix header id
+
 ## v0.2.0
 
 - [markdown-it-anchor](https://github.com/valeriangalliat/markdown-it-anchor)